requirements.txt
pandas==2.2.3
pillow==11.3.0
opencv-python==4.11.0.86
numpy==2.2.6
pytesseract==0.3.13
streamlit==1.45.1
tqdm==4.67.1
transformers==4.53.0
torch==2.7.1